Update latemarx.txt
The Project Gutenberg EBook of Manifesto of the Communist Party, by 
Karl Marx and Frederick Engels

This eBook is for the use of anyone anywhere at no cost and with
almost no restrictions whatsoever.  You may copy it, give it away or
re-use it under the terms of the Project Gutenberg License included
with this eBook or online at www.gutenberg.org
The Project Gutenberg eBook of The Eighteenth Brumaire of Louis Bonaparte, by Karl Marx

This eBook is for the use of anyone anywhere in the United States and
most other parts of the world at no cost and with almost no restrictions
whatsoever. You may copy it, give it away or re-use it under the terms
of the Project Gutenberg License included with this eBook or online at
www.gutenberg.org. If you are not located in the United States, you
will have to check the laws of the country where you are located before
using this eBook.

Title: The Eighteenth Brumaire of Louis Bonaparte

Author: Karl Marx

Translator: Daniel de Leon

Release Date: June, 1998 [eBook #1346]
[Most recently updated: April 19, 2021]

Language: English

Character set encoding: UTF-8

Produced by: An Anonymous Volunteer and David Widge